Add secondary mouse button click
I ran into issue with LGUI_T key. This didn't work, and there's an issue
on GitHub:
https://github.com/qmk/qmk_firmware/issues/6234

Correct way to get it working is to reset EEPROM, but I decided to go
with a workaround, since in my use case it doesn't matter whether it is
LGUI or RGUI is pressed.